retrieval: Change return type of retrieve_content()
retrieve_content() now returns list of (filename, s3 object key)
tuples instead of a single tuple corresponding to the chunks.

Only for CSVs, retrieve_content_csv() is called which performs
chunking, with each chunk limited to CSV_CHUNK_BYTES. No chunking
is performed for JSON and XLSX.

invoke_parser() is called in a loop with the s3 object keys.