retrieval: Store temporary files in EFS
This should allow ingestion of large files. The files on EFS are
deleted on upload to S3. This commit adds a tempdir parameter
to retrieval functions which specifies the temporary folder
where the files are saved. By default tempdir=/mnt/efs, but
can be overridden to /tmp, for example in retrieval_test.