Use correct header offset #1560
The Czechia parser was receiving a header with a spurious extra 'd'
in the second line which broke the parser.

This was happening due to incorrect calculation of the header
offset. Previously the header offset was calculated _after_
decoding, which is fine for ASCII or UTF-8, but in other encodings
there may be a BOM or in the Czechia case, additional characters
like \ufeff which made the offset off by one.

This commit correctly calculates the header offset at the byte
level before decoding.