retrieval: Fix last chunk truncation
The last chunk could have some portion unwritten, so the if block
after the loop writes the unwritten chunk (introduced in commit
8104dddd30b6078fb56418d790ed2a2f0d09a7b2). However the file
was being opened in 'w' (write) mode, which meant that the previous
contents of the file were being lost. Changed it to 'a' (append)
to fix this issue.

Reported-by: @anyalindstrombattle
(Last line of Republic of Korea dataset was not being retrieved)