retrieval: Fix last chunk truncation
The last chunk could have some portion unwritten, so the if block
after the loop writes the unwritten chunk (introduced in commit
416a3b05eedbad88d9b79d96355641938cfbed4c). However the file
was being opened in 'w' (write) mode, which meant that the previous
contents of the file were being lost. Changed it to 'a' (append)
to fix this issue.

Reported-by: @anyalindstrombattle
(Last line of Republic of Korea dataset was not being retrieved)