ingestion: Fix reauthentication during finalize_upload()
This updates the fix in 1b0659347624eae79692ebb1baf2dd488c9ef7d6

In a nested try-except the outer exception is always called, but
we do not want to raise an exception for a 401.

This changes the return type of common_lib.finalize_upload() to a
tuple of (status, text) that is checked in parsing_lib which
raises a RuntimeError if required.