ingestion: Fix reauthentication during finalize_upload()
This updates the fix in 5ee5f66a9a06958577b93f4f9831292190ef2dba

In a nested try-except the outer exception is always called, but
we do not want to raise an exception for a 401.

This changes the return type of common_lib.finalize_upload() to a
tuple of (status, text) that is checked in parsing_lib which
raises a RuntimeError if required.