data-service: Attempted fix for duplicate caserevisions insertion
batchUpsert for UUID sources is randomly failing with the error
that duplicate caserevision (same case ID and revision) is being
inserted into the caserevisions collection (issue #2005, but
seen in other UUID sources as well).

The responsibility for making sure that a case with the same
revision is not upserted again is in batchUpsertDropUnchangedCases()
which drops unchanged cases, by using findCasesWithCaseReferenceData().
As findCasesWithCaseReferenceData() calls select({}) which
only returns _id (if fieldsToSelect is not specified), this would not
be able to match against an entire case (existingCase.equalsJSON(c))

This commit drops .select() if fieldsToSelected is undefined
which should include all fields, allowing matches against existing
cases to succeed.