data-service: Add collation to sourceId index
MongoDB requires indices to have the same collation as the query
for string indices [1]. As queries in the data service use
en_US collation, all string indices in the schema require the
collation to be set.

This fixes the issue of sourceId searches taking too long that was
introduced in 5b629387b1b225126019b0a7fef61177700fde38.

[1]: https://docs.mongodb.com/v4.4/indexes/#indexes-and-collation