data-service: Ignore case revision insert errors
Case revision inserts fail intermittently when trying to
write a case revision with the same ID and revision number to
the caserevisions collection which enforces uniqueness on
these two fields. An exception is raised even if one insert
fails.

As case revisions are not exposed to the UI, and we save the
original data files for each source (which can be used to recreate
older versions of cases), this commit ignores the exception but
logs it.

This only affects UUID sources; non-UUID sources are always
ingested in full and do not update case revisions.

This is the second attempted fix for #2005, previous attempt
at 57bc84fdf9aff0434ca97a329f5866c4d9bf727e did not fix.