prune-uploads: No-op when no pending uploads for UUID sources
Currently, find_acceptable_upload() returns a tuple ([], [])
when no pending uploads are present for UUID sources.
Change this to None to bring in sync with non-UUID sources.

This should no-op prune-uploads when no pending uploads are
found. Also add a check that sources were actually ingested
in this prune before calling hooks for country export, both
in prune-uploads and in the country export hook.