Fix data-service CI failures
Data service CI was failing due to the notes removal change. After this,
setting the 'notes' field of a case document doesn't actually result in
a updated case, as the notes field is deleted by controllers, and it is
not returned from the data service APIs.

To simulate an updated document, this adds a pathogen key to a case.