Let the curator API return a null username #2672
At some point, the Google OAuth provider has changed from not
returning a name at all (i.e. undefined) to returning null for
a name. Because this can be stored in the database, it produces
500 errors when viewing the user table (even though the curator
UI could correctly handle any falsey value for the name).

This change documents that the name is nullable, allowing the
null to be passed through and correctly interpreted by the UI.